+++ src/test/bundler_test.ts
@@ -159,8 +159,9 @@ suite('Bundler', () => {
   });
 
   test('svg is nested correctly', async () => {
+    const opts = {inlineScripts: false, inlineCss: false};
     const svg =
-        dom5.query(await bundle(inputPath), matchers.template)!['content']
+        dom5.query(await bundle(inputPath, opts), matchers.template)!['content']
             .childNodes[1];
     assert.equal(svg.childNodes!.filter(dom5.isElement).length, 6);
   });
@@ -219,9 +220,10 @@ suite('Bundler', () => {
     assert.equal(bodyContainer, divActual);
   });
 
-  test('Scripts are not inlined by default', async () => {
+  test('Scripts are not inlined if specified', async () => {
     const scripts = dom5.queryAll(
-        await bundle('test/html/external.html'), matchers.externalJavascript);
+        await bundle('test/html/external.html', {inlineScripts: false}),
+        matchers.externalJavascript);
     assert.isAbove(scripts.length, 0, 'scripts were inlined');
     scripts.forEach(function(s) {
       assert.equal(dom5.getTextContent(s), '', 'script src should be empty');
@@ -248,7 +250,8 @@ suite('Bundler', () => {
   suite('Script Ordering', () => {
 
     test('Imports and scripts are ordered correctly', async () => {
-      const doc = await bundle('test/html/order-test.html');
+      const doc =
+          await bundle('test/html/order-test.html', {inlineScripts: false});
 
       const expectedOrder = [
         'first-script',